refactor: Extract session and live capture views (Issue #3 Phase 3)
Extract session management and live capture rendering functions
into dedicated modules.

Changes:
- Created session.rs (192 lines):
  * render_session_manager - Session view router
  * render_session_list - Session list display
  * render_session_save - Save session dialog
  * render_session_loading - Loading indicator
- Created live.rs (300 lines):
  * render_live_capture - Live capture view router
  * render_live_capture_header - Status header
  * render_live_packet_list - Real-time packet display
  * render_live_statistics - Live statistics dashboard
- Updated core.rs with new module imports

File size reduction: 3,393 → 2,941 lines in core (13.3% reduction)
Cumulative: 5,020 → 2,941 lines (41.4% reduction)
Total: 6,289 lines across 8 modules

Testing:
✅ cargo check --package ubertooth-cli
✅ cargo clippy --package ubertooth-cli
